+++ lib/progress_loader_overlay.dart
@@ -2,3 +2,4 @@ library progress_loader_overlay;
 
 export 'src/progress_loader.dart';
 export 'src/will_pop_scope_loader.dart';
+export 'src/pop_scope_loader.dart';